Working Jetty (experimental).
To try out the embedded Jetty server before the initial release, you
need to build the Warbler Embedded Jetty support. Change to the
`ext/jetty` directory and type `mvn install`. This creates a single
jar file with all of the Jetty dependencies as well as a main class to
launch the embedded server. That jar file is then embedded in the
resulting executable war created by Warbler.