Add config option to ignore GEM_HOME
Previous behavior was to respect GEM_HOME at runtime, so that gems
in GEM_HOME would be loaded in preference to versions packaged in
the jar file. With this commit, that behavior is still the default.
However, by setting

    config.ignore_gem_home = true

the value of GEM_HOME is not respected at runtime. This means that
gems will ONLY be loaded from within the jar file.