improve the `config.move_jars_to_webinf_lib` filter to match the default added jars
allows for quite "strangely" looking but very valid use-cases such as moving the jruby core jars upper in the class-path hierarchy (e.g. for faster re-deploys and less leakage) or if someone uses a custom - patched jruby-rack jar due a bug or a use-case of their own.

it's not expected to be used by default, yet if the filtering is there it might as well be applied to all jars we move by default to WEB-INF/lib