Fix error requiring `bundler/shared_helpers`
Currently warbler is unusable from jruby 9.2 during to this require
failing (not sure to which extent though, but at least one spec is
failing and there's an open issue about this).

If we stop monkeypatching bundler, we don't need to require it, and thus
the error disappears. Since doing that doesn't make any tests fail, I
will assume the problems caused by not monkeypatching bundler are less
important than the problems caused by doing it.

So, I'm removing the code to fix the issue and get specs green.