Use rubygems 2.7.11 in CI
On Github Actions, jruby-head installation is more standard, so it comes
with a very new bundler installed as a default gem. This means that
unless we use a rubygems version that can switch the version of bundler
according to what's locked in Gemfile.lock, things won't work because
the default version of bundler (bundler 2) will be used and that doesn't
work yet.

So, in preparation for migrating to Github Actions, I'm updating
rubygems to the minimal version that includes this feature and hope that
it works fine otherwise.