Put public folders inside public for executable war
When running `java -jar project.jar -S`, the extraction process:

1. Creates a `tmp` folder, deleted on process exit
2. Extracts files and folders from the `jar` to the `tmp` folder
  1. The layout in `tmp` is different than that of the `war`
  2. It's an attempt to pull back all the files and folders back in
     their original layout: how they were layed out before running
     `warble runnable war` at the root of the project.
3. …

This patch addresses step 2.

Previously, only files at the root were pulled back to `/public`,
leaving behind other directories at the root of `/tmp`.

This is an issue because of `Rails.public_path`.

When running the `war` inside a web container like `tomcat`:

    `Rails.public_path` = `war` root

When running the `war` with `java -jar`:

    `Rails.public_path` = `/tmp/jruby…extract/public`

So if you try to access resources from:

    `Rails.public_path / 'some' / 'nested' / 'path' / 'here.file'`

You're not going to find it because it's in:

    `Rails.root / 'some' / 'nested' / 'path' / 'here.file'`

This commit pulls back everything that is not `META-INF` back into
public.

While this is —theoretically– the inverse operation of the bundling
process, its soundness lives in code and not in data.

The optimal solution would be to bundle in the war a file listing all
transformations that happened and reversing them properly when needed
for `java -jar` execution.