Remove Bundler source hack
This hack appears to have been added around the same time we
started to support bundling dependencies from source repos, and
may no longer be necessary. My reading of the code is that it was
done to work around a bug in bundler when loading itself from a
source repo. Removing this hack allows many more tests to pass on
JRuby 10 and Bundler 2.6.3.