Update files bundler includes
I'm not positive whether this is a kosher change, given that the
spec seems to want bundler's lib files to also be included in the
jar, but it may be excluding those files because it is a default
gem. In any case, the exe files are included so we'll expect that
they are there.